Remove unused components and move page components to correct folders.
