+++ src/app/blogs/AnimatedBlogPage.tsx
@@ -63,35 +63,6 @@ export function AnimatedBlogPage() {
         transition={transitionStyle}
       />
 
-      {/* <motion.div
-      className='absolute right-80 bottom-20'
-      initial={{ rotate: 15, x: 5, y: 10 }}
-      animate={{ rotate: 5, x: 0, y: 30 }}
-      transition={transitionStyle}
-      onAnimationComplete={() => console.log('animation done')}
-      >
-        <svg
-          width="224"
-          height="137"
-          viewBox="0 0 224 137"
-          fill="none"
-          xmlns="http://www.w3.org/2000/svg"
-        >
-          <rect
-            x="23.2563"
-            width="211.689"
-            height="71.7458"
-            rx="20"
-            transform="rotate(18.9141 23.2563 0)"
-            fill="white"
-          />
-          <path
-            d="M51.0469 23.3779L46.577 21.8462L47.8509 18.1284L61.0193 22.6406L59.7454 26.3584L55.2897 24.8317L50.1454 39.8448L45.9026 38.391L51.0469 23.3779ZM59.737 37.6451C59.6227 38.0711 59.6288 38.4854 59.755 38.8881C59.8813 39.2908 60.1202 39.658 60.4716 39.9898C60.8357 40.3154 61.3063 40.5771 61.8834 40.7748C62.4321 40.9629 62.9434 41.0747 63.4175 41.1102C63.901 41.1491 64.3328 41.1385 64.7127 41.0784C65.1022 41.0216 65.4176 40.9341 65.6589 40.8159L66.4192 44.0574C66.0952 44.2318 65.6735 44.3674 65.154 44.4642C64.6473 44.5549 64.018 44.5559 63.2661 44.4674C62.5142 44.3789 61.6037 44.1515 60.5347 43.7852C59.2482 43.3444 58.1862 42.7109 57.3487 41.8849C56.5113 41.0588 55.9645 40.0786 55.7083 38.9444C55.4521 37.8101 55.5558 36.5666 56.0194 35.2138C56.4084 34.0786 57.0044 33.1411 57.8075 32.4015C58.6233 31.6557 59.595 31.18 60.7225 30.9744C61.8533 30.7593 63.0856 30.8803 64.4195 31.3374C65.6871 31.7717 66.7051 32.3795 67.4733 33.1607C68.251 33.9452 68.7334 34.8821 68.9203 35.9715C69.12 37.0547 68.9864 38.2775 68.5196 39.6397C68.4937 39.7154 68.4371 39.8651 68.3499 40.0889C68.2721 40.316 68.2107 40.4641 68.1659 40.5333L59.737 37.6451ZM64.9232 36.901C65.0175 36.5951 65.0392 36.2695 64.9883 35.9244C64.9407 35.5698 64.7979 35.2461 64.56 34.9531C64.3254 34.6507 63.9715 34.4185 63.4985 34.2564C63.0255 34.0944 62.5911 34.0512 62.1951 34.127C61.8119 34.1965 61.4849 34.3487 61.2141 34.5836C60.9434 34.8185 60.7418 35.0983 60.6094 35.4229L64.9232 36.901ZM76.1153 45.3343C76.6545 45.5191 77.1299 45.5816 77.5415 45.5218C77.9564 45.4525 78.2447 45.3822 78.4066 45.3108L79.3774 48.9098C79.1839 48.9809 78.8687 49.0526 78.4316 49.1248C77.9979 49.1876 77.4776 49.1943 76.8708 49.145C76.2735 49.0989 75.6201 48.9543 74.9106 48.7112C73.5768 48.2541 72.4896 47.6015 71.6492 46.7532C70.812 45.8956 70.2684 44.9059 70.0185 43.7843C69.778 42.666 69.8717 41.4825 70.2996 40.2338C70.7307 38.9756 71.3826 37.9833 72.2551 37.2569C73.1372 36.5338 74.1702 36.0949 75.3542 35.9404C76.5415 35.7764 77.8021 35.9229 79.1359 36.38C79.836 36.6198 80.4344 36.9095 80.9312 37.2488C81.4407 37.582 81.8523 37.908 82.1659 38.2268C82.4827 38.5362 82.6941 38.7831 82.8 38.9674L79.7737 41.2285C79.7302 41.1396 79.6396 41.0029 79.502 40.8183C79.3739 40.637 79.1888 40.452 78.9469 40.2634C78.705 40.0748 78.3948 39.9156 78.0164 39.786C77.5529 39.6272 77.0772 39.5963 76.5894 39.6934C76.111 39.7938 75.6775 40.0099 75.2889 40.3419C74.9004 40.6739 74.6137 41.1095 74.4289 41.6487C74.2441 42.1879 74.2018 42.7125 74.3019 43.2225C74.4052 43.723 74.6151 44.1596 74.9313 44.5322C75.2571 44.9081 75.6517 45.1755 76.1153 45.3343ZM93.0644 41.1526C93.8779 41.4313 94.6159 41.8692 95.2781 42.4661C95.9404 43.063 96.3991 43.8068 96.6541 44.6976C96.9124 45.5789 96.8421 46.6013 96.4434 47.7649L94.0074 54.8742L89.6937 53.396L91.8865 46.9963C92.1718 46.1638 92.1965 45.4746 91.9607 44.9287C91.7377 44.3766 91.3045 43.9903 90.6612 43.7699C90.2355 43.624 89.7898 43.5982 89.3242 43.6923C88.8617 43.777 88.4379 43.9806 88.0525 44.3031C87.6799 44.6194 87.4044 45.0377 87.2261 45.558L85.0819 51.8158L80.7539 50.3328L87.4444 30.8073L91.7723 32.2902L88.4855 41.8827C88.6965 41.6062 89.0352 41.3735 89.5017 41.1844C89.9808 40.9892 90.5304 40.8816 91.1504 40.8614C91.7831 40.8351 92.4211 40.9322 93.0644 41.1526ZM117.41 41.5825C118.574 41.9812 119.547 42.4574 120.33 43.0111C121.113 43.5647 121.72 44.0742 122.153 44.5395C122.585 45.0048 122.842 45.3043 122.923 45.4377L119.708 48.1574C119.596 47.992 119.392 47.7373 119.098 47.3933C118.813 47.0526 118.449 46.7111 118.006 46.3691C117.572 46.0302 117.086 45.7684 116.547 45.5837C115.818 45.3341 115.182 45.2639 114.637 45.3732C114.105 45.4763 113.751 45.7832 113.576 46.2941C113.456 46.6441 113.466 47.0015 113.605 47.3664C113.757 47.725 114.035 48.1163 114.44 48.5403C114.857 48.958 115.398 49.4448 116.064 50.0006C116.676 50.4959 117.23 51.0239 117.725 51.5847C118.22 52.1454 118.617 52.7464 118.915 53.3877C119.226 54.0228 119.412 54.6995 119.472 55.4178C119.532 56.1362 119.426 56.8927 119.154 57.6874C118.83 58.6334 118.354 59.3899 117.726 59.9571C117.102 60.5148 116.376 60.9056 115.549 61.1295C114.731 61.3567 113.859 61.4332 112.933 61.359C112.013 61.2976 111.099 61.1113 110.191 60.8001C108.905 60.3592 107.801 59.8172 106.88 59.1738C105.968 58.5337 105.258 57.9364 104.75 57.3819C104.243 56.8274 103.947 56.4723 103.863 56.3167L107.198 53.4794C107.282 53.6035 107.452 53.8152 107.709 54.1146C107.976 54.4173 108.31 54.7537 108.711 55.1239C109.116 55.4846 109.559 55.8267 110.04 56.1501C110.531 56.4768 111.036 56.7293 111.557 56.9075C112.427 57.2058 113.136 57.2639 113.685 57.0819C114.237 56.8904 114.607 56.5204 114.795 55.9717C114.934 55.5649 114.922 55.1537 114.758 54.738C114.594 54.3224 114.286 53.8787 113.834 53.4069C113.386 52.9256 112.806 52.3831 112.093 51.7795C111.383 51.1664 110.75 50.5001 110.193 49.7808C109.646 49.0646 109.271 48.2757 109.07 47.4138C108.868 46.552 108.947 45.596 109.307 44.5459C109.667 43.4959 110.271 42.6882 111.121 42.1229C111.982 41.5514 112.967 41.2228 114.075 41.1372C115.192 41.0548 116.304 41.2033 117.41 41.5825ZM120.087 63.8103L124.079 52.1602L128.393 53.6384L124.401 65.2884L120.087 63.8103ZM127.114 50.5206C126.433 50.2872 125.937 49.853 125.626 49.2179C125.318 48.5734 125.278 47.92 125.505 47.2578C125.732 46.5956 126.164 46.1043 126.803 45.7839C127.454 45.4573 128.115 45.4091 128.787 45.6392C129.231 45.7916 129.599 46.0445 129.89 46.398C130.185 46.742 130.382 47.1373 130.483 47.5841C130.584 48.0308 130.556 48.4812 130.4 48.9353C130.174 49.5975 129.736 50.0872 129.088 50.4043C128.444 50.712 127.786 50.7508 127.114 50.5206ZM136.703 56.1055C137.469 56.368 138.147 56.6741 138.735 57.0237C139.327 57.3639 139.808 57.6873 140.179 57.994C140.559 58.3039 140.792 58.532 140.88 58.6782L138.604 60.7369C138.47 60.5428 138.193 60.2575 137.772 59.8808C137.361 59.5074 136.857 59.2185 136.261 59.0143C135.807 58.8587 135.4 58.7984 135.039 58.8333C134.681 58.8588 134.452 59.0182 134.352 59.3114C134.251 59.6047 134.341 59.8995 134.62 60.196C134.911 60.4862 135.361 60.8094 135.968 61.1655C136.497 61.4735 137.012 61.8509 137.515 62.2979C138.017 62.7448 138.379 63.276 138.602 63.8913C138.837 64.5005 138.819 65.2024 138.546 65.997C138.267 66.8106 137.862 67.4223 137.331 67.8321C136.799 68.2419 136.195 68.4946 135.517 68.5901C134.84 68.6857 134.146 68.6751 133.435 68.5585C132.721 68.4513 132.047 68.2892 131.414 68.072C130.524 67.7673 129.776 67.4211 129.169 67.0333C128.571 66.6488 128.105 66.2829 127.769 65.9353C127.434 65.5878 127.207 65.325 127.088 65.147L129.451 62.7854C129.626 63.0778 129.947 63.4521 130.414 63.9082C130.885 64.3549 131.432 64.6852 132.057 64.8992C132.662 65.1066 133.161 65.1772 133.554 65.1109C133.947 65.0446 134.193 64.8648 134.294 64.5716C134.375 64.3351 134.359 64.1183 134.247 63.9213C134.135 63.7243 133.95 63.5235 133.692 63.319C133.435 63.1144 133.112 62.8983 132.725 62.6705C132.373 62.4652 132.011 62.2092 131.641 61.9025C131.27 61.5958 130.943 61.2404 130.658 60.8364C130.377 60.423 130.191 59.9629 130.1 59.4561C130.01 58.9494 130.068 58.3933 130.276 57.7879C130.542 57.0121 131.013 56.4548 131.689 56.1157C132.365 55.7767 133.145 55.6156 134.027 55.6324C134.912 55.6398 135.804 55.7975 136.703 56.1055ZM142.372 58.4284L144.358 59.1091L145.997 54.327L150.268 55.7906L148.63 60.5726L151.241 61.4673L150.025 65.0148L147.414 64.1202L146.301 67.3697C146.135 67.8521 146.072 68.2691 146.11 68.6204C146.151 68.9623 146.361 69.1981 146.74 69.3277C146.995 69.4152 147.232 69.433 147.451 69.3811C147.669 69.3292 147.804 69.2908 147.855 69.2659L148.28 72.8366C148.195 72.8707 147.951 72.9034 147.549 72.9349C147.157 72.9695 146.674 72.9576 146.103 72.899C145.531 72.8405 144.918 72.6994 144.266 72.4757C143.102 72.077 142.254 71.4216 141.72 70.5096C141.191 69.5881 141.161 68.4414 141.631 67.0697L143.143 62.6566L141.156 61.9759L142.372 58.4284ZM154.799 70.2184C154.685 70.6444 154.691 71.0587 154.818 71.4614C154.944 71.8641 155.183 72.2313 155.534 72.5632C155.898 72.8888 156.369 73.1504 156.946 73.3482C157.495 73.5362 158.006 73.648 158.48 73.6836C158.963 73.7224 159.395 73.7118 159.775 73.6517C160.165 73.5949 160.48 73.5074 160.721 73.3892L161.482 76.6307C161.158 76.8051 160.736 76.9407 160.217 77.0376C159.71 77.1282 159.08 77.1293 158.329 77.0408C157.577 76.9522 156.666 76.7248 155.597 76.3586C154.311 75.9177 153.249 75.2843 152.411 74.4582C151.574 73.6321 151.027 72.652 150.771 71.5177C150.515 70.3834 150.618 69.1399 151.082 67.7871C151.471 66.6519 152.067 65.7145 152.87 64.9749C153.686 64.229 154.657 63.7533 155.785 63.5477C156.916 63.3326 158.148 63.4536 159.482 63.9107C160.75 64.345 161.768 64.9528 162.536 65.734C163.314 66.5185 163.796 67.4554 163.983 68.5448C164.182 69.6281 164.049 70.8508 163.582 72.213C163.556 72.2887 163.5 72.4385 163.412 72.6623C163.335 72.8893 163.273 73.0374 163.228 73.1066L154.799 70.2184ZM159.986 69.4744C160.08 69.1684 160.102 68.8429 160.051 68.4977C160.003 68.1431 159.86 67.8194 159.622 67.5265C159.388 67.2241 159.034 66.9918 158.561 66.8298C158.088 66.6677 157.654 66.6245 157.258 66.7003C156.874 66.7698 156.547 66.922 156.277 67.1569C156.006 67.3919 155.804 67.6716 155.672 67.9962L159.986 69.4744ZM167.763 80.1465L163.449 78.6684L167.441 67.0183L171.755 68.4964L171.137 70.2986L171.066 70.2743C171.216 70.0826 171.495 69.8715 171.901 69.6409C172.311 69.4008 172.827 69.2391 173.447 69.1557C174.071 69.0629 174.766 69.1478 175.532 69.4103C176.071 69.5951 176.539 69.8504 176.935 70.1763C177.334 70.4927 177.611 70.7623 177.768 70.9849L174.892 73.9161C174.815 73.7099 174.635 73.4633 174.353 73.1764C174.083 72.8832 173.698 72.6507 173.196 72.4789C172.619 72.2811 172.091 72.2641 171.613 72.4277C171.143 72.5945 170.747 72.8553 170.425 73.2099C170.112 73.5678 169.892 73.9312 169.766 74.3002L167.763 80.1465ZM184.207 72.3826C184.973 72.6452 185.65 72.9513 186.239 73.3009C186.83 73.6411 187.311 73.9645 187.682 74.2712C188.062 74.5811 188.296 74.8092 188.383 74.9554L186.108 77.014C185.974 76.82 185.696 76.5346 185.276 76.158C184.864 75.7845 184.361 75.4957 183.765 75.2915C183.311 75.1359 182.903 75.0756 182.543 75.1105C182.185 75.136 181.956 75.2953 181.855 75.5886C181.755 75.8819 181.844 76.1767 182.123 76.4731C182.415 76.7634 182.864 77.0865 183.472 77.4426C184 77.7506 184.516 78.1281 185.018 78.575C185.52 79.022 185.883 79.5531 186.105 80.1685C186.341 80.7776 186.322 81.4795 186.05 82.2742C185.771 83.0877 185.366 83.6994 184.834 84.1093C184.303 84.5191 183.698 84.7718 183.021 84.8673C182.343 84.9628 181.649 84.9523 180.938 84.8357C180.225 84.7285 179.551 84.5663 178.917 84.3491C178.028 84.0444 177.28 83.6982 176.673 83.3105C176.075 82.926 175.608 82.56 175.273 82.2125C174.937 81.865 174.71 81.6022 174.592 81.4241L176.955 79.0626C177.129 79.355 177.45 79.7293 177.918 80.1854C178.388 80.6321 178.936 80.9624 179.56 81.1763C180.166 81.3838 180.665 81.4544 181.057 81.3881C181.45 81.3218 181.697 81.142 181.797 80.8487C181.878 80.6122 181.863 80.3955 181.75 80.1985C181.638 80.0015 181.453 79.8007 181.196 79.5961C180.938 79.3916 180.616 79.1754 180.229 78.9477C179.876 78.7423 179.515 78.4863 179.144 78.1796C178.774 77.8729 178.446 77.5176 178.162 77.1136C177.881 76.7001 177.695 76.24 177.604 75.7333C177.513 75.2266 177.572 74.6705 177.779 74.065C178.045 73.2893 178.516 72.7319 179.193 72.3929C179.869 72.0538 180.648 71.8927 181.53 71.9096C182.416 71.917 183.308 72.0747 184.207 72.3826ZM85.3615 60.5796L91.7896 62.7823C92.8586 63.1485 93.7969 63.6339 94.6046 64.2384C95.4218 64.846 95.9943 65.5813 96.3223 66.4443C96.663 67.301 96.6389 68.2969 96.2499 69.4321C95.9938 70.1795 95.6234 70.7978 95.1386 71.287C94.6633 71.7796 94.1291 72.1356 93.5358 72.3552C92.9553 72.5685 92.3744 72.6443 91.7931 72.5825C92.2466 72.8014 92.6584 73.0957 93.0285 73.4657C93.4113 73.8294 93.7175 74.262 93.9474 74.7636C94.1899 75.259 94.3196 75.8214 94.3365 76.4508C94.366 77.0741 94.2512 77.7641 93.9918 78.5209C93.632 79.571 93.0779 80.417 92.3293 81.059C91.5935 81.6949 90.6984 82.0858 89.6441 82.232C88.6026 82.3719 87.4337 82.2198 86.1377 81.7757L78.9433 79.3106L85.3615 60.5796ZM84.3483 77.3729L86.2781 78.0342C86.8836 78.2417 87.4063 78.3204 87.8463 78.2703C88.299 78.214 88.6735 78.0463 88.97 77.7674C89.2696 77.4789 89.5021 77.0935 89.6674 76.611C89.9268 75.8542 89.8799 75.2198 89.5268 74.7077C89.1865 74.1894 88.5953 73.7859 87.7534 73.4975L85.8945 72.8605L84.3483 77.3729ZM87.0322 69.54L88.7492 70.1284C89.2506 70.3002 89.6976 70.3529 90.0903 70.2866C90.4925 70.2236 90.8338 70.0604 91.1143 69.7971C91.4075 69.5276 91.627 69.18 91.7729 68.7543C92.0355 67.988 91.9945 67.3979 91.6501 66.9839C91.309 66.5604 90.76 66.2191 90.0032 65.9597L88.4423 65.4249L87.0322 69.54ZM99.0108 86.1867L94.6828 84.7037L101.373 65.1782L105.701 66.6611L99.0108 86.1867ZM107.607 89.5127C106.33 89.0751 105.284 88.426 104.469 87.5653C103.666 86.6984 103.139 85.7089 102.885 84.5968C102.645 83.4785 102.729 82.3233 103.137 81.1313C103.546 79.9394 104.188 78.9755 105.064 78.2396C105.952 77.4976 106.981 77.0255 108.149 76.8234C109.327 76.6245 110.555 76.7439 111.832 77.1815C113.109 77.6191 114.143 78.2745 114.932 79.1476C115.722 80.0207 116.231 81.0195 116.458 82.144C116.689 83.2591 116.601 84.4127 116.192 85.6046C115.784 86.7966 115.146 87.7621 114.28 88.5012C113.417 89.2308 112.407 89.6936 111.251 89.8895C110.099 90.0759 108.884 89.9503 107.607 89.5127ZM108.822 85.9652C109.314 86.1337 109.777 86.1708 110.211 86.0764C110.645 85.982 111.03 85.7702 111.365 85.4409C111.699 85.1117 111.953 84.6964 112.124 84.195C112.3 83.6841 112.356 83.196 112.293 82.7306C112.231 82.2652 112.059 81.8574 111.777 81.5072C111.495 81.157 111.108 80.8976 110.617 80.729C110.125 80.5605 109.655 80.5265 109.209 80.6271C108.771 80.731 108.38 80.9459 108.036 81.2719C107.701 81.6012 107.447 82.0212 107.271 82.5321C107.1 83.0335 107.045 83.5168 107.108 83.9822C107.18 84.4509 107.358 84.8556 107.643 85.1963C107.937 85.5403 108.33 85.7966 108.822 85.9652ZM120.435 100.488C119.621 100.21 118.837 99.8616 118.081 99.4442C117.335 99.03 116.675 98.5343 116.101 97.9571C115.527 97.3799 115.106 96.7123 114.837 95.9542L118.263 94.1946C118.393 94.5245 118.572 94.866 118.8 95.2189C119.037 95.5751 119.36 95.913 119.768 96.2327C120.186 96.5556 120.731 96.8321 121.403 97.0622C122.027 97.2762 122.608 97.3538 123.146 97.295C123.684 97.2363 124.152 97.0425 124.55 96.7138C124.958 96.3883 125.264 95.9276 125.468 95.3316L125.565 95.0478L129.808 96.5016L129.623 97.0408C129.183 98.3274 128.49 99.3004 127.545 99.9597C126.61 100.622 125.527 100.991 124.296 101.066C123.065 101.141 121.778 100.949 120.435 100.488ZM125.492 95.2607L126.139 93.3734C125.989 93.5334 125.677 93.7332 125.204 93.9727C124.741 94.2155 124.175 94.3863 123.507 94.4851C122.842 94.5744 122.131 94.4894 121.374 94.2301C120.239 93.8411 119.34 93.2264 118.676 92.386C118.016 91.5361 117.608 90.5666 117.453 89.4774C117.301 88.3788 117.419 87.2619 117.808 86.1267C118.197 84.9915 118.787 84.0414 119.578 83.2763C120.372 82.5019 121.29 81.9814 122.333 81.715C123.375 81.4486 124.464 81.5099 125.599 81.8989C126.328 82.1485 126.926 82.4855 127.393 82.91C127.864 83.3251 128.215 83.7518 128.445 84.1901C128.679 84.619 128.817 84.9886 128.858 85.2988L129.442 83.596L133.727 85.0644L129.735 96.7145L125.492 95.2607ZM122.037 87.5756C121.855 88.1054 121.815 88.6253 121.915 89.1352C122.018 89.6358 122.233 90.074 122.558 90.4498C122.894 90.829 123.312 91.1044 123.813 91.2762C124.334 91.4545 124.833 91.4935 125.311 91.3931C125.79 91.2928 126.214 91.0734 126.583 90.7349C126.966 90.3903 127.248 89.953 127.429 89.4233C127.611 88.8935 127.654 88.38 127.561 87.8827C127.479 87.3792 127.281 86.9411 126.964 86.5685C126.651 86.1864 126.235 85.9062 125.714 85.7279C125.213 85.5561 124.712 85.5219 124.212 85.6252C123.724 85.7223 123.284 85.9416 122.892 86.283C122.503 86.615 122.218 87.0459 122.037 87.5756Z"
-            fill="#45084A"
-          />
-        </svg>
-      </motion.div> */}
-
       {/* tech siters blog text pad */}
       <motion.rect
         x="270.256"
