+++ src/app/landing-page/Gallery.tsx
@@ -3,7 +3,7 @@
 import React from 'react';
 
 import Image from 'next/image';
-import Button from './Button';
+import Button from '../../components/Button';
 
 const Gallery = () => {
   // Array of gallery images
