+++ src/app/meet-the-team/page.tsx
@@ -1,11 +1,7 @@
 'use client';
 
 import React from 'react';
-
-//import { Card } from "@/components/ui/card";
-import TeamMember from '@/components/TeamMember';
-//import imaniGrace from "./assets/imani_Grace.svg";
-//import EclipseSvg from "@/components/svgs/eclise_svg";
+import TeamMember from '@/app/meet-the-team/TeamMember';
 
 const teamMembers = [
   {
@@ -13,7 +9,6 @@ const teamMembers = [
     role: 'Founder',
     imageUrl: './imanigrace.svg',
     bio: "At the heart of TSK is a passionate software developer whose love for building extends far beyond just lines of code. She's a baker, a dreamer, and a firm believer in the power of community. Her journey into tech wasn't just about finding a career—it was about creating space for women like her to feel seen, supported, and celebrated.\n\nWith one hand on her keyboard and the other in a mixing bowl, she blends logic with creativity, tech with tenderness. Baking has always been her way of expressing love, and building TSK became her way of extending that love to the tech world—especially to women navigating it.\n\nThrough TSK, she's built more than a community—she's sparked a movement. One where women support women, share knowledge, lift each other up, and grow together in a space where everyone belongs.",
-    /*componentImage : <EclipseSvg /> */
   },
   {
     name: 'Imani Grace',
@@ -91,7 +86,6 @@ const MeetTheTeam = () => {
               role={member.role}
               imageUrl={member.imageUrl}
               bio={member.bio}
-              //componentImage={member.componentImage}
             />
           ))}
         </div>
