--- src/components/AnimatedShapes.tsx
@@ -1,208 +1,0 @@
-'use client';
-
-import React from 'react';
-import Image from 'next/image';
-import techSister1 from '@/assets/Group 633181.svg';
-import techSister2 from '@/assets/IMG_4912 1.svg';
-
-const AnimatedShapes: React.FC = () => {
-  return (
-    <div className="relative w-full h-screen flex flex-col items-center justify-center overflow-hidden">
-      <div className="relative flex flex-col items-center gap-4 z-10 w-full max-w-4xl">
-        <div className="relative w-1/2 h-40 max-w-[16rem]">
-          <Image
-            src={techSister1}
-            alt="Tech Sister 1"
-            width={180}
-            height={180}
-            className="absolute z-10 -bottom-0 left-1/4"
-          />
-        </div>
-
-        <div className="relative w-72 h-48">
-          <Image
-            src={techSister2}
-            width={220}
-            height={220}
-            alt="Tech Sister 2"
-            className="relative object-contain z-10 bottom-10 left-[20%]"
-          />
-        </div>
-
-        <div>
-          <svg
-            width="539"
-            height="166"
-            viewBox="0 0 539 166"
-            fill="none"
-            xmlns="http://www.w3.org/2000/svg"
-          >
-            <rect y="0.0712891" width="539" height="165.929" rx="82.9643" fill="#45084A" />
-          </svg>
-        </div>
-      </div>
-
-      {/* semi-circle Shape */}
-      <div className="absolute left-[5%] top-[15%] w-32 h-32 md:left-[10%] md:top-[25%] animate-elasticswing">
-        <svg
-          width="181"
-          height="189"
-          viewBox="0 0 181 189"
-          fill="none"
-          xmlns="http://www.w3.org/2000/svg"
-        >
-          <path
-            d="M152.464 168.216C112.204 200.681 53.2494 194.364 20.7839 154.104C-11.6816 113.845 -5.36359 54.8903 34.8955 22.4248L62.7035 9.10996e-05L180.271 145.791L152.464 168.216Z"
-            fill="#45084A"
-          />
-        </svg>
-      </div>
-      {/* Circle Shape */}
-      <div className="absolute left-[10%] bottom-[25%] w-20 h-20  rounded-full opacity-75 md:left-[20%] md:bottom-[30%] animate-zoom">
-        <svg
-          width="120"
-          height="120"
-          viewBox="0 0 120 120"
-          fill="none"
-          xmlns="http://www.w3.org/2000/svg"
-        >
-          <circle cx="60" cy="60" r="60" fill="#FFBAFF" />
-        </svg>
-      </div>
-      {/* Cloud Shape */}
-      <div className="absolute left-[8%] top-[0%] w-32 h-20 rounded-[100px] md:left-[10%] md:top-[2%] animate-elasticswing">
-        <svg
-          width="96"
-          height="79"
-          viewBox="0 0 96 79"
-          fill="none"
-          xmlns="http://www.w3.org/2000/svg"
-        >
-          <path
-            d="M16.9636 10.3243C25.0148 5.6404 34.3337 5.54951 41.517 9.28225L41.9536 9.5084L42.1879 9.07608C43.4975 6.65115 45.5351 4.48393 48.1769 2.94698C55.2994 -1.19641 64.0652 0.738763 67.7857 7.13381C68.6439 8.6091 69.1625 10.2021 69.3714 11.8151L69.4512 12.4319L70.0365 12.2224C78.7777 9.08234 88.372 12.0047 92.7653 19.5562C97.7815 28.1791 94.2023 39.705 84.6347 45.2713L33.7916 74.8504C23.0079 81.1239 9.69703 78.2111 4.036 68.4807C-0.437543 60.7912 1.14388 51.0728 7.31477 44.2291L7.56949 43.9465L7.36496 43.626C7.15256 43.2928 6.94675 42.9692 6.75503 42.6397C0.442338 31.7889 4.95023 17.3133 16.9636 10.3243Z"
-            fill="#45084A"
-            stroke="#70169E"
-          />
-        </svg>
-      </div>
-      {/* Code Bracket */}
-      <div className="absolute right-[8%] top-[28%] text-tsk-primary-dark text-6xl md:right-[12%] md:top-[40%] z-10">
-        <svg
-          width="54"
-          height="44"
-          viewBox="0 0 54 44"
-          fill="none"
-          xmlns="http://www.w3.org/2000/svg"
-        >
-          <g clipPath="url(#clip0_2657_1390)">
-            <path
-              d="M33.1423 0.100934C31.708 -0.312503 30.2145 0.522809 29.8011 1.95718L19.0011 39.7572C18.5877 41.1916 19.423 42.685 20.8573 43.0984C22.2917 43.5119 23.7852 42.6766 24.1986 41.2422L34.9986 3.44218C35.412 2.00781 34.5767 0.514372 33.1423 0.100934ZM39.943 10.2344C38.8883 11.2891 38.8883 13.0019 39.943 14.0566L47.4777 21.5997L39.9345 29.1428C38.8798 30.1975 38.8798 31.9103 39.9345 32.965C40.9892 34.0197 42.702 34.0197 43.7567 32.965L53.2067 23.515C54.2614 22.4603 54.2614 20.7475 53.2067 19.6928L43.7567 10.2428C42.702 9.18812 40.9892 9.18812 39.9345 10.2428L39.943 10.2344ZM14.0652 10.2344C13.0105 9.17968 11.2977 9.17968 10.243 10.2344L0.792969 19.6844C-0.261719 20.7391 -0.261719 22.4519 0.792969 23.5066L10.243 32.9566C11.2977 34.0112 13.0105 34.0112 14.0652 32.9566C15.1198 31.9019 15.1198 30.1891 14.0652 29.1344L6.52203 21.5997L14.0652 14.0566C15.1198 13.0019 15.1198 11.2891 14.0652 10.2344Z"
-              fill="#45084A"
-            />
-          </g>
-          <defs>
-            <clipPath id="clip0_2657_1390">
-              <rect width="54" height="43.2" fill="white" />
-            </clipPath>
-          </defs>
-        </svg>
-      </div>
-      {/* Pill Shape */}
-      <div className="absolute right-[10%] top-[15%] w-12 h-32 rounded-full md:right-[14%] md:top-[10%] animate-squashstretch">
-        <svg
-          width="70"
-          height="144"
-          viewBox="0 0 70 144"
-          fill="none"
-          xmlns="http://www.w3.org/2000/svg"
-        >
-          <rect
-            x="10"
-            y="134"
-            width="124"
-            height="50"
-            rx="25"
-            transform="rotate(-90 10 134)"
-            stroke="#70169E"
-            strokeWidth="20"
-          />
-        </svg>
-      </div>
-      {/* Texts */}
-      <div className="absolute right-[20%] top-[2%] bg-foreground px-4 py-2 rounded-full md:right-[25%] md:top-[4%] animate-elasticswing">
-        <svg
-          width="126"
-          height="56"
-          viewBox="0 0 126 56"
-          fill="none"
-          xmlns="http://www.w3.org/2000/svg"
-        >
-          <rect
-            x="-2"
-            y="18.3174"
-            width="126"
-            height="40.8847"
-            rx="20"
-            transform="rotate(-9.27922 -2 18.3174)"
-            fill="white"
-          />
-        </svg>
-      </div>
-      <div className="absolute right-[3%] top-[40%] bg-foreground px-4 py-2 rounded-full md:right-[8%] md:top-[45%] animate-elasticswing">
-        <svg
-          width="224"
-          height="137"
-          viewBox="0 0 224 137"
-          fill="none"
-          xmlns="http://www.w3.org/2000/svg"
-        >
-          <rect
-            x="23.2563"
-            width="211.689"
-            height="71.7458"
-            rx="20"
-            transform="rotate(18.9141 23.2563 0)"
-            fill="white"
-          />
-          <path
-            d="M51.0469 23.3779L46.577 21.8462L47.8509 18.1284L61.0193 22.6406L59.7454 26.3584L55.2897 24.8317L50.1454 39.8448L45.9026 38.391L51.0469 23.3779ZM59.737 37.6451C59.6227 38.0711 59.6288 38.4854 59.755 38.8881C59.8813 39.2908 60.1202 39.658 60.4716 39.9898C60.8357 40.3154 61.3063 40.5771 61.8834 40.7748C62.4321 40.9629 62.9434 41.0747 63.4175 41.1102C63.901 41.1491 64.3328 41.1385 64.7127 41.0784C65.1022 41.0216 65.4176 40.9341 65.6589 40.8159L66.4192 44.0574C66.0952 44.2318 65.6735 44.3674 65.154 44.4642C64.6473 44.5549 64.018 44.5559 63.2661 44.4674C62.5142 44.3789 61.6037 44.1515 60.5347 43.7852C59.2482 43.3444 58.1862 42.7109 57.3487 41.8849C56.5113 41.0588 55.9645 40.0786 55.7083 38.9444C55.4521 37.8101 55.5558 36.5666 56.0194 35.2138C56.4084 34.0786 57.0044 33.1411 57.8075 32.4015C58.6233 31.6557 59.595 31.18 60.7225 30.9744C61.8533 30.7593 63.0856 30.8803 64.4195 31.3374C65.6871 31.7717 66.7051 32.3795 67.4733 33.1607C68.251 33.9452 68.7334 34.8821 68.9203 35.9715C69.12 37.0547 68.9864 38.2775 68.5196 39.6397C68.4937 39.7154 68.4371 39.8651 68.3499 40.0889C68.2721 40.316 68.2107 40.4641 68.1659 40.5333L59.737 37.6451ZM64.9232 36.901C65.0175 36.5951 65.0392 36.2695 64.9883 35.9244C64.9407 35.5698 64.7979 35.2461 64.56 34.9531C64.3254 34.6507 63.9715 34.4185 63.4985 34.2564C63.0255 34.0944 62.5911 34.0512 62.1951 34.127C61.8119 34.1965 61.4849 34.3487 61.2141 34.5836C60.9434 34.8185 60.7418 35.0983 60.6094 35.4229L64.9232 36.901ZM76.1153 45.3343C76.6545 45.5191 77.1299 45.5816 77.5415 45.5218C77.9564 45.4525 78.2447 45.3822 78.4066 45.3108L79.3774 48.9098C79.1839 48.9809 78.8687 49.0526 78.4316 49.1248C77.9979 49.1876 77.4776 49.1943 76.8708 49.145C76.2735 49.0989 75.6201 48.9543 74.9106 48.7112C73.5768 48.2541 72.4896 47.6015 71.6492 46.7532C70.812 45.8956 70.2684 44.9059 70.0185 43.7843C69.778 42.666 69.8717 41.4825 70.2996 40.2338C70.7307 38.9756 71.3826 37.9833 72.2551 37.2569C73.1372 36.5338 74.1702 36.0949 75.3542 35.9404C76.5415 35.7764 77.8021 35.9229 79.1359 36.38C79.836 36.6198 80.4344 36.9095 80.9312 37.2488C81.4407 37.582 81.8523 37.908 82.1659 38.2268C82.4827 38.5362 82.6941 38.7831 82.8 38.9674L79.7737 41.2285C79.7302 41.1396 79.6396 41.0029 79.502 40.8183C79.3739 40.637 79.1888 40.452 78.9469 40.2634C78.705 40.0748 78.3948 39.9156 78.0164 39.786C77.5529 39.6272 77.0772 39.5963 76.5894 39.6934C76.111 39.7938 75.6775 40.0099 75.2889 40.3419C74.9004 40.6739 74.6137 41.1095 74.4289 41.6487C74.2441 42.1879 74.2018 42.7125 74.3019 43.2225C74.4052 43.723 74.6151 44.1596 74.9313 44.5322C75.2571 44.9081 75.6517 45.1755 76.1153 45.3343ZM93.0644 41.1526C93.8779 41.4313 94.6159 41.8692 95.2781 42.4661C95.9404 43.063 96.3991 43.8068 96.6541 44.6976C96.9124 45.5789 96.8421 46.6013 96.4434 47.7649L94.0074 54.8742L89.6937 53.396L91.8865 46.9963C92.1718 46.1638 92.1965 45.4746 91.9607 44.9287C91.7377 44.3766 91.3045 43.9903 90.6612 43.7699C90.2355 43.624 89.7898 43.5982 89.3242 43.6923C88.8617 43.777 88.4379 43.9806 88.0525 44.3031C87.6799 44.6194 87.4044 45.0377 87.2261 45.558L85.0819 51.8158L80.7539 50.3328L87.4444 30.8073L91.7723 32.2902L88.4855 41.8827C88.6965 41.6062 89.0352 41.3735 89.5017 41.1844C89.9808 40.9892 90.5304 40.8816 91.1504 40.8614C91.7831 40.8351 92.4211 40.9322 93.0644 41.1526ZM117.41 41.5825C118.574 41.9812 119.547 42.4574 120.33 43.0111C121.113 43.5647 121.72 44.0742 122.153 44.5395C122.585 45.0048 122.842 45.3043 122.923 45.4377L119.708 48.1574C119.596 47.992 119.392 47.7373 119.098 47.3933C118.813 47.0526 118.449 46.7111 118.006 46.3691C117.572 46.0302 117.086 45.7684 116.547 45.5837C115.818 45.3341 115.182 45.2639 114.637 45.3732C114.105 45.4763 113.751 45.7832 113.576 46.2941C113.456 46.6441 113.466 47.0015 113.605 47.3664C113.757 47.725 114.035 48.1163 114.44 48.5403C114.857 48.958 115.398 49.4448 116.064 50.0006C116.676 50.4959 117.23 51.0239 117.725 51.5847C118.22 52.1454 118.617 52.7464 118.915 53.3877C119.226 54.0228 119.412 54.6995 119.472 55.4178C119.532 56.1362 119.426 56.8927 119.154 57.6874C118.83 58.6334 118.354 59.3899 117.726 59.9571C117.102 60.5148 116.376 60.9056 115.549 61.1295C114.731 61.3567 113.859 61.4332 112.933 61.359C112.013 61.2976 111.099 61.1113 110.191 60.8001C108.905 60.3592 107.801 59.8172 106.88 59.1738C105.968 58.5337 105.258 57.9364 104.75 57.3819C104.243 56.8274 103.947 56.4723 103.863 56.3167L107.198 53.4794C107.282 53.6035 107.452 53.8152 107.709 54.1146C107.976 54.4173 108.31 54.7537 108.711 55.1239C109.116 55.4846 109.559 55.8267 110.04 56.1501C110.531 56.4768 111.036 56.7293 111.557 56.9075C112.427 57.2058 113.136 57.2639 113.685 57.0819C114.237 56.8904 114.607 56.5204 114.795 55.9717C114.934 55.5649 114.922 55.1537 114.758 54.738C114.594 54.3224 114.286 53.8787 113.834 53.4069C113.386 52.9256 112.806 52.3831 112.093 51.7795C111.383 51.1664 110.75 50.5001 110.193 49.7808C109.646 49.0646 109.271 48.2757 109.07 47.4138C108.868 46.552 108.947 45.596 109.307 44.5459C109.667 43.4959 110.271 42.6882 111.121 42.1229C111.982 41.5514 112.967 41.2228 114.075 41.1372C115.192 41.0548 116.304 41.2033 117.41 41.5825ZM120.087 63.8103L124.079 52.1602L128.393 53.6384L124.401 65.2884L120.087 63.8103ZM127.114 50.5206C126.433 50.2872 125.937 49.853 125.626 49.2179C125.318 48.5734 125.278 47.92 125.505 47.2578C125.732 46.5956 126.164 46.1043 126.803 45.7839C127.454 45.4573 128.115 45.4091 128.787 45.6392C129.231 45.7916 129.599 46.0445 129.89 46.398C130.185 46.742 130.382 47.1373 130.483 47.5841C130.584 48.0308 130.556 48.4812 130.4 48.9353C130.174 49.5975 129.736 50.0872 129.088 50.4043C128.444 50.712 127.786 50.7508 127.114 50.5206ZM136.703 56.1055C137.469 56.368 138.147 56.6741 138.735 57.0237C139.327 57.3639 139.808 57.6873 140.179 57.994C140.559 58.3039 140.792 58.532 140.88 58.6782L138.604 60.7369C138.47 60.5428 138.193 60.2575 137.772 59.8808C137.361 59.5074 136.857 59.2185 136.261 59.0143C135.807 58.8587 135.4 58.7984 135.039 58.8333C134.681 58.8588 134.452 59.0182 134.352 59.3114C134.251 59.6047 134.341 59.8995 134.62 60.196C134.911 60.4862 135.361 60.8094 135.968 61.1655C136.497 61.4735 137.012 61.8509 137.515 62.2979C138.017 62.7448 138.379 63.276 138.602 63.8913C138.837 64.5005 138.819 65.2024 138.546 65.997C138.267 66.8106 137.862 67.4223 137.331 67.8321C136.799 68.2419 136.195 68.4946 135.517 68.5901C134.84 68.6857 134.146 68.6751 133.435 68.5585C132.721 68.4513 132.047 68.2892 131.414 68.072C130.524 67.7673 129.776 67.4211 129.169 67.0333C128.571 66.6488 128.105 66.2829 127.769 65.9353C127.434 65.5878 127.207 65.325 127.088 65.147L129.451 62.7854C129.626 63.0778 129.947 63.4521 130.414 63.9082C130.885 64.3549 131.432 64.6852 132.057 64.8992C132.662 65.1066 133.161 65.1772 133.554 65.1109C133.947 65.0446 134.193 64.8648 134.294 64.5716C134.375 64.3351 134.359 64.1183 134.247 63.9213C134.135 63.7243 133.95 63.5235 133.692 63.319C133.435 63.1144 133.112 62.8983 132.725 62.6705C132.373 62.4652 132.011 62.2092 131.641 61.9025C131.27 61.5958 130.943 61.2404 130.658 60.8364C130.377 60.423 130.191 59.9629 130.1 59.4561C130.01 58.9494 130.068 58.3933 130.276 57.7879C130.542 57.0121 131.013 56.4548 131.689 56.1157C132.365 55.7767 133.145 55.6156 134.027 55.6324C134.912 55.6398 135.804 55.7975 136.703 56.1055ZM142.372 58.4284L144.358 59.1091L145.997 54.327L150.268 55.7906L148.63 60.5726L151.241 61.4673L150.025 65.0148L147.414 64.1202L146.301 67.3697C146.135 67.8521 146.072 68.2691 146.11 68.6204C146.151 68.9623 146.361 69.1981 146.74 69.3277C146.995 69.4152 147.232 69.433 147.451 69.3811C147.669 69.3292 147.804 69.2908 147.855 69.2659L148.28 72.8366C148.195 72.8707 147.951 72.9034 147.549 72.9349C147.157 72.9695 146.674 72.9576 146.103 72.899C145.531 72.8405 144.918 72.6994 144.266 72.4757C143.102 72.077 142.254 71.4216 141.72 70.5096C141.191 69.5881 141.161 68.4414 141.631 67.0697L143.143 62.6566L141.156 61.9759L142.372 58.4284ZM154.799 70.2184C154.685 70.6444 154.691 71.0587 154.818 71.4614C154.944 71.8641 155.183 72.2313 155.534 72.5632C155.898 72.8888 156.369 73.1504 156.946 73.3482C157.495 73.5362 158.006 73.648 158.48 73.6836C158.963 73.7224 159.395 73.7118 159.775 73.6517C160.165 73.5949 160.48 73.5074 160.721 73.3892L161.482 76.6307C161.158 76.8051 160.736 76.9407 160.217 77.0376C159.71 77.1282 159.08 77.1293 158.329 77.0408C157.577 76.9522 156.666 76.7248 155.597 76.3586C154.311 75.9177 153.249 75.2843 152.411 74.4582C151.574 73.6321 151.027 72.652 150.771 71.5177C150.515 70.3834 150.618 69.1399 151.082 67.7871C151.471 66.6519 152.067 65.7145 152.87 64.9749C153.686 64.229 154.657 63.7533 155.785 63.5477C156.916 63.3326 158.148 63.4536 159.482 63.9107C160.75 64.345 161.768 64.9528 162.536 65.734C163.314 66.5185 163.796 67.4554 163.983 68.5448C164.182 69.6281 164.049 70.8508 163.582 72.213C163.556 72.2887 163.5 72.4385 163.412 72.6623C163.335 72.8893 163.273 73.0374 163.228 73.1066L154.799 70.2184ZM159.986 69.4744C160.08 69.1684 160.102 68.8429 160.051 68.4977C160.003 68.1431 159.86 67.8194 159.622 67.5265C159.388 67.2241 159.034 66.9918 158.561 66.8298C158.088 66.6677 157.654 66.6245 157.258 66.7003C156.874 66.7698 156.547 66.922 156.277 67.1569C156.006 67.3919 155.804 67.6716 155.672 67.9962L159.986 69.4744ZM167.763 80.1465L163.449 78.6684L167.441 67.0183L171.755 68.4964L171.137 70.2986L171.066 70.2743C171.216 70.0826 171.495 69.8715 171.901 69.6409C172.311 69.4008 172.827 69.2391 173.447 69.1557C174.071 69.0629 174.766 69.1478 175.532 69.4103C176.071 69.5951 176.539 69.8504 176.935 70.1763C177.334 70.4927 177.611 70.7623 177.768 70.9849L174.892 73.9161C174.815 73.7099 174.635 73.4633 174.353 73.1764C174.083 72.8832 173.698 72.6507 173.196 72.4789C172.619 72.2811 172.091 72.2641 171.613 72.4277C171.143 72.5945 170.747 72.8553 170.425 73.2099C170.112 73.5678 169.892 73.9312 169.766 74.3002L167.763 80.1465ZM184.207 72.3826C184.973 72.6452 185.65 72.9513 186.239 73.3009C186.83 73.6411 187.311 73.9645 187.682 74.2712C188.062 74.5811 188.296 74.8092 188.383 74.9554L186.108 77.014C185.974 76.82 185.696 76.5346 185.276 76.158C184.864 75.7845 184.361 75.4957 183.765 75.2915C183.311 75.1359 182.903 75.0756 182.543 75.1105C182.185 75.136 181.956 75.2953 181.855 75.5886C181.755 75.8819 181.844 76.1767 182.123 76.4731C182.415 76.7634 182.864 77.0865 183.472 77.4426C184 77.7506 184.516 78.1281 185.018 78.575C185.52 79.022 185.883 79.5531 186.105 80.1685C186.341 80.7776 186.322 81.4795 186.05 82.2742C185.771 83.0877 185.366 83.6994 184.834 84.1093C184.303 84.5191 183.698 84.7718 183.021 84.8673C182.343 84.9628 181.649 84.9523 180.938 84.8357C180.225 84.7285 179.551 84.5663 178.917 84.3491C178.028 84.0444 177.28 83.6982 176.673 83.3105C176.075 82.926 175.608 82.56 175.273 82.2125C174.937 81.865 174.71 81.6022 174.592 81.4241L176.955 79.0626C177.129 79.355 177.45 79.7293 177.918 80.1854C178.388 80.6321 178.936 80.9624 179.56 81.1763C180.166 81.3838 180.665 81.4544 181.057 81.3881C181.45 81.3218 181.697 81.142 181.797 80.8487C181.878 80.6122 181.863 80.3955 181.75 80.1985C181.638 80.0015 181.453 79.8007 181.196 79.5961C180.938 79.3916 180.616 79.1754 180.229 78.9477C179.876 78.7423 179.515 78.4863 179.144 78.1796C178.774 77.8729 178.446 77.5176 178.162 77.1136C177.881 76.7001 177.695 76.24 177.604 75.7333C177.513 75.2266 177.572 74.6705 177.779 74.065C178.045 73.2893 178.516 72.7319 179.193 72.3929C179.869 72.0538 180.648 71.8927 181.53 71.9096C182.416 71.917 183.308 72.0747 184.207 72.3826ZM78.639 68.3924L82.4464 80.5109L77.3805 78.7751L74.4249 68.6133L71.6194 76.801L67.3765 75.3472L73.7947 56.6163L78.0376 58.0701L75.4119 65.7327L83.3446 59.8886L88.1267 61.5271L78.639 68.3924ZM88.4363 77.0771C88.3221 77.5031 88.3281 77.9174 88.4544 78.3201C88.5807 78.7228 88.8195 79.09 89.1709 79.4219C89.5351 79.7475 90.0056 80.0091 90.5827 80.2069C91.1314 80.3949 91.6428 80.5067 92.1168 80.5423C92.6003 80.5811 93.0321 80.5705 93.4121 80.5104C93.8015 80.4536 94.1169 80.3661 94.3583 80.2479L95.1186 83.4894C94.7946 83.6638 94.3728 83.7994 93.8534 83.8963C93.3466 83.9869 92.7173 83.988 91.9654 83.8994C91.2135 83.8109 90.3031 83.5835 89.2341 83.2172C87.9475 82.7764 86.8855 82.1429 86.0481 81.3169C85.2106 80.4908 84.6638 79.5107 84.4076 78.3764C84.1515 77.2421 84.2552 75.9986 84.7187 74.6458C85.1077 73.5106 85.7037 72.5732 86.5068 71.8336C87.3227 71.0877 88.2943 70.612 89.4219 70.4064C90.5526 70.1913 91.7849 70.3123 93.1188 70.7694C94.3865 71.2037 95.4044 71.8115 96.1727 72.5927C96.9504 73.3772 97.4327 74.3141 97.6197 75.4035C97.8193 76.4867 97.6858 77.7095 97.219 79.0717C97.1931 79.1474 97.1365 79.2971 97.0492 79.5209C96.9714 79.748 96.9101 79.8961 96.8652 79.9653L88.4363 77.0771ZM93.6225 76.3331C93.7168 76.0271 93.7385 75.7016 93.6876 75.3564C93.64 75.0018 93.4972 74.6781 93.2593 74.3852C93.0247 74.0828 92.6709 73.8505 92.1979 73.6885C91.7249 73.5264 91.2904 73.4832 90.8945 73.559C90.5112 73.6285 90.1842 73.7807 89.9135 74.0156C89.6427 74.2505 89.4411 74.5303 89.3088 74.8549L93.6225 76.3331ZM109.581 76.41C110.394 76.6888 111.132 77.1266 111.794 77.7236C112.457 78.3205 112.915 79.0643 113.17 79.9551C113.429 80.8364 113.358 81.8588 112.96 83.0224L110.524 90.1316L106.196 88.6486L108.389 82.2489C108.674 81.4164 108.689 80.724 108.434 80.1716C108.183 79.6098 107.736 79.2186 107.092 78.9982C106.667 78.8523 106.226 78.8281 105.769 78.9255C105.326 79.0166 104.916 79.2251 104.54 79.5508C104.177 79.8704 103.906 80.2903 103.728 80.8106L101.584 87.0684L97.2702 85.5903L101.262 73.9402L105.576 75.4183L104.988 77.1353C105.208 76.8621 105.551 76.6309 106.018 76.4419C106.497 76.2467 107.047 76.139 107.667 76.1189C108.299 76.0926 108.937 76.1896 109.581 76.41ZM120.184 89.2716L126.047 82.4329L130.276 83.8819L116.258 98.6927L111.944 97.2146L116.663 92.2353L115.788 78.9175L120.087 80.3908L120.184 89.2716ZM136.675 99.0925L137.322 97.2053C137.172 97.3653 136.86 97.5651 136.387 97.8046C135.924 98.0474 135.358 98.2182 134.69 98.317C134.025 98.4063 133.314 98.3213 132.557 98.0619C131.422 97.673 130.523 97.0583 129.859 96.2178C129.199 95.368 128.791 94.3984 128.636 93.3093C128.484 92.2107 128.602 91.0938 128.991 89.9585C129.38 88.8233 129.97 87.8732 130.761 87.1082C131.555 86.3337 132.473 85.8133 133.516 85.5469C134.558 85.2805 135.647 85.3418 136.782 85.7308C137.511 85.9804 138.109 86.3174 138.576 86.7419C139.047 87.157 139.398 87.5837 139.628 88.022C139.862 88.4509 140 88.8204 140.041 89.1307L140.625 87.4279L144.91 88.8963L140.918 100.546L136.675 99.0925ZM133.22 91.4075C133.038 91.9373 132.998 92.4571 133.098 92.9671C133.201 93.4676 133.416 93.9058 133.741 94.2817C134.077 94.6609 134.495 94.9363 134.996 95.1081C135.517 95.2864 136.016 95.3254 136.494 95.225C136.973 95.1246 137.397 94.9052 137.766 94.5668C138.149 94.2221 138.431 93.7849 138.612 93.2552C138.794 92.7254 138.837 92.2119 138.744 91.7146C138.662 91.2111 138.464 90.773 138.147 90.4004C137.834 90.0183 137.418 89.7381 136.897 89.5598C136.396 89.388 135.895 89.3538 135.395 89.4571C134.907 89.5542 134.467 89.7735 134.075 90.1149C133.686 90.4469 133.401 90.8777 133.22 91.4075Z"
-            fill="#45084A"
-          />
-        </svg>
-      </div>
-      {/* Logo */}
-      <div className="absolute right-[0%] bottom-[20%] w-16 h-16 sm:w-16 sm:h-16 md:w-18 md:h-18 md:right-[0%] md:bottom-[25%] animate-zoom">
-        <svg
-          width="63"
-          height="84"
-          viewBox="0 0 63 84"
-          fill="none"
-          xmlns="http://www.w3.org/2000/svg"
-        >
-          <path
-            d="M46.3729 62.3133C44.7709 62.3133 43.4722 60.7973 43.4722 58.9304C43.4722 57.0634 44.7709 55.5498 46.3729 55.5498C47.9748 55.5498 49.2712 57.0634 49.2712 58.9304C49.2712 60.7973 47.9724 62.3133 46.3729 62.3133Z"
-            fill="#70169E"
-          />
-          <path
-            d="M0.0787022 24.8401C0.988304 19.1915 8.7235 12.342 12.4622 9.20976C14.1596 7.78925 15.5658 6.68866 16.3918 6.05122C16.4993 5.88649 16.7619 5.45914 16.6043 5.19653C16.5231 5.06045 16.2868 4.90049 15.4512 5.02464C12.1088 5.52599 8.52773 7.88952 8.52773 7.88952C7.33403 8.68214 5.7846 9.75409 4.03702 11.1101C4.74608 9.69918 7.94998 3.69008 15.1599 1.09974C24.8337 -2.36438 33.1443 3.41314 33.9058 3.95985C31.9548 4.59007 30.0417 5.33219 28.1761 6.18253C25.5499 7.37623 22.1574 9.18588 22.2983 9.52489C22.4582 9.90449 27.0731 8.20705 34.8465 6.66001C38.2175 5.99153 40.9869 5.58806 42.9469 7.13749C43.2191 7.35236 44.5417 8.39804 44.5346 9.83764C44.5346 11.0672 43.5557 11.972 43.1021 12.3802C40.3518 14.8727 38.9695 16.1213 33.9058 18.8883C31.571 20.1632 28.3957 22.362 25.1703 26.3537C30.0355 24.3359 34.7123 21.8912 39.1462 19.0483C42.5196 16.8733 45.967 14.3069 48.6767 15.5531C50.0232 16.1738 50.6034 17.463 50.7418 17.7782C51.8973 20.3637 50.6392 23.9854 47.3995 26.6712C49.2855 25.7831 51.3649 25.9717 52.6518 27.1487C53.8455 28.2589 53.9075 29.8489 53.9219 30.324C54.0054 32.9692 52 34.6834 51.6992 34.9317C51.6299 35.2349 51.5774 35.6479 51.7875 35.8198C52.1074 36.0824 52.7425 35.5596 53.6378 35.328C54.6357 35.0677 55.9249 35.1823 56.7223 35.7577C57.7942 36.536 58.1547 38.2836 57.6462 40.0741C57.7484 40.3909 57.872 40.7004 58.0163 41.0004C58.711 42.4329 59.4296 42.6716 59.9262 43.4666C60.4228 44.2616 60.2318 45.8708 59.802 49.0174C59.5919 50.5525 59.4773 50.8437 59.6182 51.4835C59.8569 52.5889 60.5732 53.4269 61.3252 54.0357C61.7902 54.4211 62.174 54.895 62.4544 55.4299C62.5261 55.9838 62.3637 56.2034 62.2682 56.2918C62.1727 56.3801 62.0557 56.447 61.2202 56.478C60.535 56.5043 60.3559 56.4589 60.2342 56.5998C60.1603 56.7235 60.1267 56.867 60.1378 57.0107C60.1489 57.1543 60.2043 57.291 60.2962 57.4019C60.6615 57.8651 61.4279 57.5165 61.7143 57.9582C61.7704 58.0624 61.7997 58.1789 61.7997 58.2972C61.7997 58.4155 61.7704 58.532 61.7143 58.6362C61.5472 58.9728 60.9981 58.6935 60.4801 59.0683C60.3941 59.1304 60.1363 59.3071 60.1745 59.4766C60.2437 59.7798 61.318 59.6437 61.3897 59.9541C61.435 60.1403 61.0793 60.3695 61.0268 60.4029C60.7164 60.6058 60.4801 60.5533 60.2151 60.6416C59.7161 60.8016 59.4702 61.3292 59.3628 61.5584C59.007 62.3056 59.3914 62.6757 59.1001 63.3728C58.8972 63.8146 58.5516 64.1753 58.1189 64.397C56.9992 64.97 55.6169 64.28 54.8291 63.8861C53.8168 63.3776 53.8216 63.0863 52.9502 62.733C52.3335 62.4857 51.676 62.3561 51.0116 62.351C51.1601 61.7974 51.2355 61.2267 51.236 60.6536C51.236 57.5237 49.0611 54.9859 46.3777 54.9859C43.6942 54.9859 41.5169 57.5237 41.5169 60.6536C41.5169 63.7835 43.6918 66.326 46.3777 66.326C46.5525 66.3258 46.7271 66.3146 46.9005 66.2926C46.5591 67.1163 46.2893 67.9328 46.0745 68.5821C45.1195 71.4972 44.0523 74.7225 45.3917 78.4899C46.1897 80.6654 47.5807 82.5741 49.4073 84H30.7855C32.7355 80.498 34.4171 76.8531 35.8158 73.0967C39.1319 64.2347 39.4852 61.1382 39.49 59.9015C39.481 56.8332 38.8567 53.7979 37.6541 50.975C36.8586 51.2575 36.045 51.4865 35.2189 51.6602C33.2827 52.0661 32.3063 52.2714 31.2057 51.9157C29.5823 51.3928 28.0352 49.767 28.2572 48.5423C28.4291 47.6064 29.6037 47.0692 29.3674 46.4915C29.2671 46.2527 28.9663 46.1358 28.7705 46.0641C28.2776 45.9171 27.7488 45.9475 27.276 46.1501C25.5888 46.7531 23.8617 47.2381 22.1073 47.6016C21.2765 47.7544 18.6384 48.2295 17.3325 46.7899C16.2677 45.6272 16.3942 43.5597 17.0746 42.1273C18.1728 39.8235 20.417 39.864 24.2918 37.8562C27.2402 36.3307 26.8964 35.815 33.1252 31.1524C34.7963 29.9062 37.3246 28.0679 40.5572 25.9407C37.9194 26.9686 35.3389 28.1377 32.8267 29.443C29.0523 31.4055 26.099 32.9406 22.7065 35.6789C20.1019 37.7775 17.1319 39.4271 14.6777 41.7C14.1668 42.1774 12.9755 43.3043 11.5168 43.1515C10.989 43.0779 10.4873 42.8759 10.0556 42.5633C9.62401 42.2507 9.27563 41.8371 9.04102 41.3586C8.44656 40.1362 8.85003 38.7849 9.21053 37.8562C12.3882 29.7128 21.7707 21.1181 21.7707 21.1181C23.8191 19.2464 26.7341 16.668 30.3987 13.7315C28.7613 13.944 27.1459 14.3011 25.5714 14.7987C20.0541 16.5415 16.2605 19.6451 14.1286 21.4166C11.2995 23.7777 7.98818 25.6542 5.62942 28.4713C5.27109 28.9253 4.77788 29.2539 4.22085 29.4096C2.85525 29.7486 1.1411 28.7984 0.420101 27.4448C0.0231362 26.6377 -0.0968488 25.7222 0.0787022 24.8401Z"
-            fill="#70169E"
-          />
-          <path
-            d="M39.9126 77.5917L36.5702 78.7854L35.9829 78.0931L37.7018 74.9775L38.2677 75.6436L36.6729 78.4154L36.3577 78.043L39.3492 76.9352L39.9126 77.5917Z"
-            fill="#FFBAFF"
-          />
-          <path
-            d="M40.9845 78.5524L38.4849 72.0419L39.2273 71.4141L41.7246 77.9245L40.9845 78.5524Z"
-            fill="#FFBAFF"
-          />
-          <path
-            d="M42.6582 75.2647L42.0924 74.5963L43.6753 71.8245L43.9904 72.1993L40.999 73.3071L40.4355 72.641L43.7898 71.4473L44.3772 72.1396L42.6582 75.2647Z"
-            fill="#FFBAFF"
-          />
-        </svg>
-      </div>
-    </div>
-  );
-};
-
-export default AnimatedShapes;
