--- src/components/svg_components/svg2.tsx
@@ -1,10 +1,0 @@
-import * as React from 'react';
-const SvgComponent = (props: React.JSX.IntrinsicAttributes & React.SVGProps<SVGSVGElement>) => (
-  <svg xmlns="http://www.w3.org/2000/svg" width={109} height={113} fill="none" {...props}>
-    <path
-      fill="#45084A"
-      d="M90.779 100.505c-23.839 19.224-58.748 15.483-77.971-8.356C-6.416 68.31-2.675 33.4 21.163 14.178L38.744 0l69.616 86.327-17.581 14.178Z"
-    />
-  </svg>
-);
-export default SvgComponent;
